Changes module name to srfi-116
In addition the two calls to (include ...) need to be referenced
relatively to the files ilists-base and ilists-impl. If not, when the
egg tries to build it will not know where to look for them, because
`chicken-install` uses the top-level directory as the current working
directory. (To avoid this just move everything out of ilists, but this
would potentially make the branches diverge too much).