Adds setup, meta, and release-info to egg
The setup file is fairly straightforward, it merely compiles the
module as located in ilists/ilists.scm and renames it to srfi-116 (this
is done so we can call the egg by it's srfi name).

The meta file provides metadata about the egg, such as what
dependencies it has, the author, where to download the repo etc.

Lastly, the release-info file helps the coop determine where to
download the egg (which tags are releases, for example). With this, a
tag called 1.0 should be made and uploaded to Github so that the egg
can be added to the coop.