Updates meta/setup/release-info as per srfi-113
See [this pull
request](https://github.com/scheme-requests-for-implementation/srfi-113/pull/5)
for a bit more context. The basic problem is that given the previous
release-info / meta files, the entire repo would be downloaded
(including html / tar.gz files) every time you install the egg. This is
unecessary as there is no need for these files for the egg to install.